Finished (hopefully) Song Library Project
Kinda overhauled the alphabetized methods and  created my own comparing since it was easier for me to compare between individual songs and artists. I also was able to maintain the indices of add and edited songs in my way. Its really messy because it was error after I fixed another but I think I got it now. Check it out and tell me if you find anything wrong?